window geometry adjustments

+++ cancelOrder.py
@@ -48,7 +48,7 @@ class CancelOrder(QWidget):
         self.screen = QApplication.primaryScreen()
         self.screenWidth = self.screen.size().width()
         self.screenHeight = self.screen.size().height()
-        self.move(self.screenWidth/2. - self.width/2., self.screenHeight/2. - self.height/2.)
+        self.move(int(self.screenWidth/2. - self.width/2.), int(self.screenHeight/2. - self.height/2.))
 
         self.setFixedWidth(self.width)
         self.setFixedHeight(self.height)
