+++ symsettings.py
@@ -227,7 +227,7 @@ class SymbolSettings(QWidget):
         self.screen = QApplication.primaryScreen()
         self.screenWidth = self.screen.size().width()
         self.screenHeight = self.screen.size().height()
-        self.move(self.screenWidth/2. - self.width/2., self.screenHeight/7.)
+        self.move(int(self.screenWidth/2. - self.width/2.), int(self.screenHeight/7.))
 
         self.setFixedWidth(self.width)
         # self.setMinimumWidth(500)
